// src/components/form/types.ts
/**
 * Form Control Types
 */
export enum FormControlTypes {
    Checkbox = 1,
    ColorPicker = 2,
    Email = 3,
    Datalist = 4,
    Dropdown = 5,
    File = 6,
    ListBox = 7,
    MultiCheckbox = 8,
    MultiDropdown = 9,
    MultiListBox = 10,
    MultiRadio = 11,
    MultiSwitch = 12,
    Password = 13,
    Radio = 14,
    Range = 15,
    Readonly = 16,
    Switch = 17,
    TextArea = 18,
    TextField = 19
}

/**
 * Form Validation Types
 */
export enum FormValidationTypes {
    Default = 1,
    Tooltip = 2
}